refactor: moved new helper functions in a new pyrdp.mitm.cli package
More to come
