+++ pyrdp/core/__init__.py
@@ -7,7 +7,7 @@
 from pyrdp.core.defer import defer
 from pyrdp.core.event import EventEngine
 from pyrdp.core.FileProxy import FileProxy, FileProxyObserver
-from pyrdp.core.helpers import decodeUTF16LE, encodeUTF16LE, parseTarget, validateKeyAndCertificate
+from pyrdp.core.helpers import decodeUTF16LE, encodeUTF16LE
 from pyrdp.core.observer import CompositeObserver, Observer
 from pyrdp.core.packing import Int16BE, Int16LE, Int32BE, Int32LE, Int8, Uint16BE, Uint16LE, Uint32BE, Uint32LE, \
     Uint64LE, Uint8
